+++ src/client/App.vue
@@ -773,7 +773,7 @@ onBeforeUnmount(() => {
         />
       </template>
       <template #footer>
-        <span>Wing 0.6.0</span>
+        <span>Wing 0.6.3</span>
         <span v-if="lastRefreshAt">最近刷新 {{ lastRefreshAt.toLocaleTimeString('zh-CN', { hour12: false }) }}</span>
       </template>
     </PnwWorkbenchShell>